move printing logs to worker thread to help avoid deadlocks
Because `uconsole_output_text` may be implemented in terms of an unreal function call, if process event is locking it may deadlock.
This happens if external code attempts to acquire a lock during a hook, while at the same time trying to print a log message on the thread which holds that lock.

While it's really the external code's responsibility to avoid this, the fact that every single log message can cause it is absolutely awful, it transitively affects half the codebase in non-obvious ways.
By moving printing to it's own worker thread, we can be sure that thread will never be holding a conflicting lock.
As a bonus, this means we move the filesystem access to the worker thread too - though we still do string formatting and copying on the logic threads.

If someone hooks output text for some reason, while it will trigger hooks on the worker thread when we print, that thread will always already hold the process event lock, so it won't be able to deadlock.