Change log all calls to write to dedicated file
When it's enabled, we generate more log messages than the standard logging queue can keep up with. If logging to console in bl2/tps it actually recurses, every log generates half a dozen more calls which it all has to log too.

Swapping to a dedicated file, and doing blocking writes helps avoid this, and also means we don't lose as much data if interrupted. Don't want to flush on every line to avoid killing performance even more.