Update call function signature for TPS
Seems this has been broken a long time.

The ghidra script which generates the signatures decided the stack offset was a perfectly sane constant to hardcode in.
While the signature didn't match CallFunction in TPS, it did still match something, so we never got any logs about a null pointer.

This fixes missing a lot of hooks - most notably the say bypass.